Revert "Merged PR 716: Batching for download telemetry
Currently we are sending all the download telemetry for content in bulk.
Added a message size to send the download telemetry in batches instead.
Once a batch fails to get sent, we add it back to the pending APIs bucket and increment the retry count.

We maintain two batches here - contentDataBatch and contentPropertiesBatch
each batch is filled till the message size and then cleared once sent out

Related work items: #3847"