Merged PR 727: Sending download telemetry logic in batches
Currently we are sending all the download telemetry data together in one request.

Making batches of 25 and sending the telemetry data instead
We send telemetry twice - to cloud and for iot central graphing, for this some more properties are sent.

If data has been sent to both cloud telemetry and iot central for graphing, we delete it from the pending API bucket list. In case there is an error in one of these, we add it back to our pending API bucket with incremented retry count.

Tested by adding more contents in **mock_server.py** file

**hub_config_tmp.ini** file is purely meant for testing
**hub_config.ini** is the actual file consumed during runtime

Related work items: #3847, #3848