Merged PR 863: Removing device sdk folder
Earlier in the Pilot, a python based IoT edge client was present. All GRPC calls to it were made using these scripts in the device SDK folder. All the GRPC calls now are independent of this hence can be removed.

Validated that the code builds and runs after removing the folder.

Related work items: #4165, #4168